Header, MainPage, Footer, HeroSection, WhyMe-Section & Contact added (only structure)
Added the structure of the Header, MainPage, Footer, HeroSection,
WhyMe-Section & Contact. No Design, no images added, only the basic
structure to display the section where they belong.

Whole App will be mobile first, so this is only the Mobile POV. Desktop
Structure will be added later via media-queries on 'any-pointer'. It
will be a combination of 'any-pointer', 'not' and 'coarse', because
Desktops do not have any input device with a coarse pointer.