Added Projects and the ability to grow and shrink with transition
Added a project component and a components for all projects. Its now
possible to render a list of projects in a certain structure. currently
its possible to grow and shrink the individual project. Width and height
of the individual projects needs some work. Also, the end result should
be: only one project should be open, the others should be collapsed.