Changed Structure of Nav, Header and Menu and Mobile Version is done with placeholder texts, etc
Mobile Version now done beside the texts and images and form validation,
form sending. CSS wise its done.
Also changed the Strcuture of in which component the header element
lies. Header is now in header component formerly the nav component. Nav
is inside the header component. The menu for links, etc is now in the
body and is still fixed with a little less z-index than the header, to
make it slide under the header when active and back when inactive.