Legal Page implemented with Legal Notice Text in German (no real styling, yet) + centered Social Links von Desktop
There is a legal page now path: "/legal" and a legal text in german.
needs to be translated in english. Also restructured a few components to
make the router more useful. end-banner on desktop now separate and the
links on the last page (desktop) also now outside of the
main-page-component to have it in all routes.